+++ babelfish/__init__.py
@@ -5,15 +5,16 @@
 # that can be found in the LICENSE file.
 #
 __title__ = 'babelfish'
-__version__ = '0.1.5'
+__version__ = '0.2.0'
 __author__ = 'Antoine Bertin'
 __license__ = 'BSD'
 __copyright__ = 'Copyright 2013 the BabelFish authors'
 
-from .language import CONVERTERS, LANGUAGES, Language, register_converter, unregister_converter, load_converters, clear_converters
-from .country import COUNTRIES, Country
 from .converters import Converter, ReverseConverter
-from .exceptions import *
+from .country import COUNTRIES, Country
+from .exceptions import Error, ConvertError, ReverseError
+from .language import (CONVERTERS, LANGUAGES, Language, register_converter, unregister_converter, load_converters,
+    clear_converters)
 
 
 load_converters()